Change IPv4 Router ID to 8 octets
On Mar 27, 2013, at 2:57 AM, Saikat Ray (sairay) wrote:

Another thing I noticed is that IPv4 (IPv6) router-id TLV size has been
changed back to 4 + 1 (16 + 1). That doesn't work for OSPFv3. In v3,
the link-state ID of the network LSA is the (4 byte) "interface ID"
(and not the interface IP address) of the DR. So unless we put the
router-id of the DR in the node descriptor, the "pseudo-node" cannot be
uniquely identified with the interface ID alone.